Remove test server binary
