+++ service/src/lib.rs
@@ -1,4 +1,2 @@
 pub mod client;
 pub mod service;
-pub mod test_server;
-pub mod tls_config;
