+++ service/src/main.rs
@@ -5,7 +5,7 @@ use eyre::Context;
 use tonic::transport::{Certificate, Identity};
 
 use post::pow::randomx::RandomXFlag;
-use post_service::{client, tls_config::Tls};
+use post_service::client;
 
 /// Post Service
 #[derive(Parser, Debug)]
@@ -99,6 +99,24 @@ enum RandomXMode {
     Fast,
     /// Light mode for verification. Requires only 256 MiB of memory, but runs significantly slower
     Light,
+}
+
+/// TLS configuration
+///
+/// Either all fields must be specified or none
+#[derive(Args, Debug, Clone)]
+#[group(required = false)]
+pub struct Tls {
+    /// CA certificate
+    #[arg(long, required = false)]
+    pub ca_cert: PathBuf,
+    #[arg(long, required = false)]
+    pub cert: PathBuf,
+    #[arg(long, required = false)]
+    pub key: PathBuf,
+    /// domain name to verify the certificate of server against
+    #[arg(long, default_value = "localhost")]
+    pub domain: String,
 }
 
 impl std::fmt::Display for RandomXMode {
